Fixed sql java mapping standards

+++ domain/src/test/java/za/co/entelect/bootcamp/twoface/squareeyes/domain/TwoFaceTests.java
@@ -32,7 +32,7 @@ public class TwoFaceTests {
     @Test
     public void issueExists() {
         System.out.print(": Issue Exists...");
-        Issues cutIssue = new Issues(new Date(), "title","publisher",001,"description");
+        Issues cutIssue = new Issues(new Date(), "title","publisher",(short)001,"description");
         Assert.assertEquals("title",cutIssue.getIssueTitle());
         System.out.print(" : test passed." + "\n");
     }
@@ -48,7 +48,7 @@ public class TwoFaceTests {
     @Test
     public void stockExists() {
         System.out.print(": Stock Exists...");
-        Stock cutStock = new Stock(01, new Issues(), "condition", 10, BigDecimal.ONE);
+        Stock cutStock = new Stock(01, new Issues(), "condition", (short)10, BigDecimal.ONE);
         Assert.assertEquals("condition",cutStock.getCondition());
         System.out.print(" : test passed." + "\n");
     }
@@ -64,7 +64,7 @@ public class TwoFaceTests {
     @Test
     public void orderExists() {
         System.out.print(": Order Exists...");
-        Orders cutOrder = new Orders(new Date(), 5, BigDecimal.valueOf(10), "shipmentRef",new Date(), "deliveryStatus",new Issues(),new Suppliers());
+        Orders cutOrder = new Orders(new Date(), (short)5, BigDecimal.valueOf(10), "shipmentRef",new Date(), "deliveryStatus",new Issues(),new Suppliers());
         Assert.assertEquals("shipmentRef",cutOrder.getShipmentRef());
         System.out.print(" : test passed." + "\n");
     }
@@ -73,7 +73,7 @@ public class TwoFaceTests {
     public void issueRetrievePublicationDate() {
         System.out.print(": Issue Exists...");
         Date temp = new Date();
-        Issues cutIssue = new Issues(temp, "title","publisher",001,"description");
+        Issues cutIssue = new Issues(temp, "title","publisher",(short)001,"description");
         Assert.assertEquals(temp,cutIssue.getPublicationDate());
         System.out.print(" : test passed." + "\n");
     }
@@ -81,9 +81,9 @@ public class TwoFaceTests {
     @Test
     public void ordersIssuesRelationshipMapping() {
         System.out.print(": Issue & Order mapping");
-        Issues tempIssue =  new Issues(new Date(), "title","publisher",001,"description");
+        Issues tempIssue =  new Issues(new Date(), "title","publisher",(short)001,"description");
         Suppliers tempSuppliers = new Suppliers("supplierName", "supplierCity", "supplierReferenceNumber");
-        Orders cutOrder = new Orders(new Date(), 5, BigDecimal.valueOf(10), "shipmentRef",new Date(), "deliveryStatus",tempIssue,tempSuppliers);
+        Orders cutOrder = new Orders(new Date(), (short)5, BigDecimal.valueOf(10), "shipmentRef",new Date(), "deliveryStatus",tempIssue,tempSuppliers);
         Assert.assertEquals(tempIssue.getID(),cutOrder.getIssue());
         System.out.print(" : test passed." + "\n");
     }
@@ -91,9 +91,9 @@ public class TwoFaceTests {
     @Test
     public void ordersSuppliersRelationshipMapping() {
         System.out.print(": Supplier & Order mapping");
-        Issues tempIssue =  new Issues(new Date(), "title","publisher",001,"description");
+        Issues tempIssue =  new Issues(new Date(), "title","publisher",(short)001,"description");
         Suppliers tempSuppliers = new Suppliers("supplierName", "supplierCity", "supplierReferenceNumber");
-        Orders cutOrder = new Orders(new Date(), 5, BigDecimal.valueOf(10), "shipmentRef",new Date(), "deliveryStatus",tempIssue,tempSuppliers);
+        Orders cutOrder = new Orders(new Date(), (short)5, BigDecimal.valueOf(10), "shipmentRef",new Date(), "deliveryStatus",tempIssue,tempSuppliers);
         Assert.assertEquals(tempSuppliers.getID(),cutOrder.getSuppliers());
         System.out.print(" : test passed." + "\n");
     }
